refact(bonusNeededFrame): remove bonusNeededFrame):
