+++ src/main/java/bowling/domain/Bonus.java
@@ -1,0 +1,11 @@
+package bowling.domain;
+
+public enum Bonus {
+    STRIKE(2), SPARE(1);
+
+    private int bonus;
+
+    Bonus(int bonus){
+        this.bonus = bonus;
+    }
+}
