+++ src/main/java/bowling/domain/Player.java
@@ -27,4 +27,5 @@ public class Player {
     public String name() {
         return name;
     }
+
 }
