+++ src/main/java/bowling/domain/frame/NormalFrame.java
@@ -1,7 +1,6 @@
 package bowling.domain.frame;
 
-import bowling.domain.score.NormalScore;
-import bowling.domain.score.Score;
+import bowling.domain.Score;
 import bowling.domain.pin.NormalPins;
 import bowling.domain.pin.Pins;
 
@@ -10,7 +9,7 @@ public class NormalFrame implements Frame {
     private NormalPins normalPins;
 
     public NormalFrame() {
-        score = new NormalScore();
+        score = new Score();
         normalPins = new NormalPins();
     }
 
@@ -22,7 +21,7 @@ public class NormalFrame implements Frame {
     @Override
     public void bowl(int pin) {
         normalPins.bowl(pin);
-        score.addScore(pin);
+//        score.addScore(pin);
     }
 
     @Override
